+++ src/Symfony/Component/FeatureToggle/Tests/FeatureCheckerTest.php
@@ -73,4 +73,70 @@ final class FeatureCheckerTest extends TestCase
         self::assertFalse($featureChecker->isEnabled('fake-2'));
         self::assertFalse($featureChecker->isEnabled('fake-3'));
     }
+
+    /**
+     * @dataProvider provideIsEnableFallback
+     */
+    public function testIsEnableFallback(bool $expected, StrategyResult $result, bool $default)
+    {
+        $feature = new Feature(
+            name: 'fake',
+            description: 'Fake description',
+            default: $default,
+            strategy: new class($result) implements StrategyInterface {
+                public function __construct(private readonly StrategyResult $result) {}
+
+                public function compute(): StrategyResult
+                {
+                    return $this->result;
+                }
+            },
+        );
+
+        $featureChecker = new FeatureChecker(new FeatureCollection([$feature]),true);
+
+        self::assertSame($expected, $featureChecker->isEnabled('fake'));
+    }
+
+    public function provideIsEnableFallback(): iterable
+    {
+        // Grant
+        yield "grant should ignore fallback #1" => [
+            true,
+            StrategyResult::Grant,
+            true,
+        ];
+
+        yield "grant should ignore fallback #2" => [
+            true,
+            StrategyResult::Grant,
+            false,
+        ];
+
+        // Deny
+        yield "deny should ignore fallback #1" => [
+            false,
+            StrategyResult::Deny,
+            true,
+        ];
+
+        yield "deny should ignore fallback #2" => [
+            false,
+            StrategyResult::Deny,
+            false,
+        ];
+
+        // Abstain
+        yield "abstain should use fallback #1" => [
+            true,
+            StrategyResult::Abstain,
+            true,
+        ];
+
+        yield "abstain should use fallback #2" => [
+            false,
+            StrategyResult::Abstain,
+            false,
+        ];
+    }
 }
